egressgw: add support for excludedCIDRs
Add support to the egress gateway manager for the new excludedCIDRs
policy property.

The implementation is based on the idea of adding, for each excluded
CIDR, "zero" entries to the egress policy map (i.e. entries with the
gateway IP set to 0.0.0.0). When traffic is then matched against these
entries in the datapath, it will skip altogether all the egress gateway
logic.

Moreover, to make IP rules and routes work with excluded CIDRs, a couple
of new policy helpers have been introduced, to allow iterating and
matching over all the endpoints and destinations CIDRs minus the
excluded ones:

* forEachEndpointAndDestinationMinusExcludedCIDRs: works like
  forEachEndpointAndDestination, but iterate over
  destinationCIDRs - excludedCIDRs
* matchesMinusExcludedCIDRs: works like matches but matches over
  destinationCIDRs - excludedCIDRs

This allow to install IP rules for all the actual destination CIDRs
space.

A practical example may be useful here: let's say a policy specifies the
1.1.1.0/24 destination CIDR with 1.1.1.1/22 excluded. To target only the
effective CIDR space the manager will install rules for the following
destination CIDRs:

	1.1.1.0/28
	1.1.1.16/30
	1.1.1.20/31
	1.1.1.23/32
	1.1.1.24/29
	1.1.1.32/27
	1.1.1.64/26
	1.1.1.128/25

which represent the diff between 1.1.1.0/24 and 1.1.1.1/22.

Fixes: #23002
Signed-off-by: Gilberto Bertin <jibi@cilium.io>